pyproject.toml
[project]
name = "patres-task-test"
version = "0.1.0"
description = "Library Management System"
authors = [
    {name = "devorasp", email = "devorasp@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "pydantic>=1.8.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "alembic>=1.7.0",
    "psycopg2-binary>=2.9.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "httpx>=0.23.0",
    "python-multipart>=0.0.5"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=app --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "**/__pycache__",
    "**/migrations/*",
    "**/tests/*"
]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

[tool.black]
line-length = 79
target-version = ["py310"]
include = "\\.pyi?$"
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 79
force_sort_within_sections = true
known_first_party = ["app"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["."], exclude = ["data", "materials", "tests"]}}
